Refactor ForthCore into ForthEngine for improved functionality
ForthCore class was deleted and its relevant parts were integrated into ForthEngine class. The restructuring was essentially done to consolidate related functionalities in a single class. This enhances code readability and maintainability. Also, the code was optimized by removing the invocation of redundant functions - `ForthWord.undefinedWord()` and `ForthWord.isNumber`. The logic of these functions has been included directly in ForthEngine. As a result, duplicate method calls are eliminated which may lead to improved performance.
