+++ java/forth/src/main/java/ForthEvaluator.java
@@ -1,10 +1,9 @@
-import forth.ForthCore;
 import forth.ForthEngine;
 
 import java.util.List;
 
 class ForthEvaluator {
-    private final ForthEngine forthEngine = new ForthCore();
+    private final ForthEngine forthEngine = new ForthEngine();
 
     public List<Integer> evaluateProgram(List<String> commands) {
         commands.forEach(forthEngine);
